requirements.txt
# Alpaca SDK
alpaca-py>=0.21.0,<1.0.0

# 設定管理
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# データ取得
yfinance>=0.2.30,<1.0.0

# LLM出力バリデーション
jsonschema>=4.20.0,<5.0.0

# 市場カレンダー
exchange-calendars>=4.5.0,<5.0.0

# テクニカル指標（Phase 2で使用）
pandas>=2.1.0,<3.0.0
numpy>=1.26.0,<2.0.0

# マクロデータ（Phase 2で使用）
fredapi>=0.5.0,<1.0.0
